requirements.txt
fastapi
uvicorn[standard]
httpx
pdfplumber
python-docx
pydantic
asyncpg
boto3
